Fix import errors from merging file moves
